services/gateway/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
aiokafka>=0.13.0
prometheus-client==0.20.0
opentelemetry-api==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-instrumentation-fastapi==0.46b0
